client/pyproject.toml
[tool.poetry]
name = "albinears"
version = "0.1.0"
description = "A sound recorder and monitor application to run mobile deviceses and enabling the client endpoint to act as sound sensing device. The application uses the kivy environment to run on android and apple devices."
authors = ["Karl Jonas Riisnaes"]
license = "No licence yet"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
Kivy = "^2.3.0"
kivymd = "^1.2.0"
pandas = "^2.2.0"
plyer = "^2.1.0"
pyobjus = "^1.2.3"
python-for-android = "^2024.1.21"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
